TestFx setup (#100) - Test FX has been fixed for correct setup of the application - 1 test running
