+++ pom.xml
@@ -66,6 +66,11 @@
             <version>3.11.1</version>
             <scope>test</scope>
         </dependency>
+<!--        <dependency>-->
+<!--            <groupId>org.loadui</groupId>-->
+<!--            <artifactId>testFx</artifactId>-->
+<!--            <version>3.1.2</version>-->
+<!--        </dependency>-->
     </dependencies>
 
     <build>
