--- tests_e2e/com/cbms/app/MainTest.java
@@ -1,42 +1,0 @@
-package com.cbms.app;
-
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
-import javafx.scene.input.KeyCode;
-import javafx.scene.input.MouseButton;
-import javafx.stage.Stage;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
-import org.testfx.api.FxToolkit;
-import org.testfx.framework.junit.ApplicationTest;
-
-import java.util.Objects;
-
-public class MainTest extends ApplicationTest {
-
-    @Override
-    public void start (Stage primaryStage) throws Exception {
-        Parent mainNode = FXMLLoader.load(Objects.requireNonNull(getClass().getClassLoader().getResource("/Systems.fxml")));
-        primaryStage.setScene(new Scene(mainNode));
-        primaryStage.show();
-        primaryStage.toFront();
-    }
-
-    @Before
-    public void setUp () throws Exception {
-    }
-
-    @After
-    public void tearDown () throws Exception {
-        FxToolkit.hideStage();
-        release(new KeyCode[]{});
-        release(new MouseButton[]{});
-    }
-
-    @Test
-    public void testAccessingAddSystemPage () {
-        clickOn("#addSystemButton");
-    }
-}
